Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main(){
-    int r, area;
+    float r, area;
     scanf("%d",&r);
     area = 3.14*r*r;
-    printf("Area: %d",area);
+    printf("Area: %.2f",area);
     return 0;
 
 }
